in firefox, webpack-dev-server generates the following warning whenever the page reloads:
"The character encoding of the HTML document was not declared. The
document will render with garbled text in some browser configurations if
the document contains characters from outside the US-ASCII range. The
character encoding of the page must to be declared in the document or in
the transfer protocol."

Adding:
<meta charset="utf-8"/>
fixes it.